add NUMA and shared-L2-cache detect code (required at work) enable most of IA-32 specific code to be used in amd64 (resides in directory lib/sysdep/x86_x64)
bits: add IsBitSet
remove mem_PageSize (use os_cpu_PageSize instead)
cpuid: change interface to allow gracefully supporting later subfunctions that require input parameters
amd64_asm.asm: add amd64 implementation of cpuid
cpu: move functions provided by OS to sysdep/os_cpu.cpp
cpu topology: avoid trouble when process affinity is set by remapping processor numbers to 0..PopulationCount(processAffinity)
topology.cpp: move ex-ia32 topology code here.

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@5945 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/sysdep/win/wposix/waio.cpp
@@ -405,7 +405,7 @@ int aio_suspend(const struct aiocb* const cbs[], int n, const struct timespec* t
 	const BOOL waitAll = FALSE;
 	// convert timespec to milliseconds (ts == 0 => no timeout)
 	const DWORD timeout = ts? (DWORD)(ts->tv_sec*1000 + ts->tv_nsec/1000000) : INFINITE;
-	DWORD result = WaitForMultipleObjects(numPendingIos, hEvents, waitAll, timeout);
+	DWORD result = WaitForMultipleObjects((DWORD)numPendingIos, hEvents, waitAll, timeout);
 
 	for(size_t i = 0; i < numPendingIos; i++)
 		ResetEvent(hEvents[i]);
